+++ libraries/stdlib/js-ir-minimal-for-test/common-src/minimalCommonRanges.kt
@@ -15,7 +15,7 @@ package kotlin.ranges
  */
 public infix fun UByte.until(to: UByte): UIntRange {
     if (to <= UByte.MIN_VALUE) return UIntRange.EMPTY
-    return this.toUInt() .. (to - 1u).toUInt()
+    return this.toUInt()..to - 1u
 }
 
 /**
@@ -25,7 +25,7 @@ public infix fun UByte.until(to: UByte): UIntRange {
  */
 public infix fun UInt.until(to: UInt): UIntRange {
     if (to <= UInt.MIN_VALUE) return UIntRange.EMPTY
-    return this .. (to - 1u).toUInt()
+    return this..to - 1u
 }
 
 /**
@@ -35,7 +35,7 @@ public infix fun UInt.until(to: UInt): UIntRange {
  */
 public infix fun ULong.until(to: ULong): ULongRange {
     if (to <= ULong.MIN_VALUE) return ULongRange.EMPTY
-    return this .. (to - 1u).toULong()
+    return this..to - 1u
 }
 
 /**
@@ -45,5 +45,5 @@ public infix fun ULong.until(to: ULong): ULongRange {
  */
 public infix fun UShort.until(to: UShort): UIntRange {
     if (to <= UShort.MIN_VALUE) return UIntRange.EMPTY
-    return this.toUInt() .. (to - 1u).toUInt()
+    return this.toUInt()..to - 1u
 }
